requirements.txt
pyserial==3.5
matplotlib==3.8.2
numpy==1.24.3
scipy==1.11.4
pygame==2.5.2
